refactor: decompose stateful execution runtime
